feat(st-15): Organized controllers and models
