+++ app/controllers/api/v1/authorize/passwords_controller.rb
@@ -1,17 +1,18 @@
 module Api
-    module V1
+  module V1
+    module Authorize
       class PasswordsController < ApplicationController
         def forgot_password
-            result = Passwords::Forgot.call(email: params[:email])
-            if result.success?
-              render json: { message: result.message }, status: :ok
-            else
-              render json: { error: result.error }, status: :unprocessable_entity
-            end
+          result = ::Passwords::Forgot.call(email: params[:email])
+          if result.success?
+            render json: { message: result.message }, status: :ok
+          else
+            render json: { error: result.error }, status: :unprocessable_entity
+          end
         end
 
         def check_otp
-          result = Passwords::CheckOtp.call(email: params[:email], otp: params[:otp])
+          result = ::Passwords::CheckOtp.call(email: params[:email], otp: params[:otp])
           if result.success?
             render json: { message: result.message }, status: :ok
           else
@@ -20,7 +21,7 @@ module Api
         end
 
         def reset_password
-          result = Passwords::Reset.call(email: params[:email], otp: params[:otp], password: params[:password])
+          result = ::Passwords::Reset.call(email: params[:email], otp: params[:otp], password: params[:password])
           if result.success?
             render json: { message: result.message }, status: :ok
           else
@@ -29,4 +30,5 @@ module Api
         end
       end
     end
+  end
 end
