+++ app/interactors/passwords/check_otp.rb
@@ -3,7 +3,7 @@ module Passwords
       include Interactor
 
       def call
-        user = User.find_by(email: context.email, otp: context.otp)
+        user = Users::User.find_by(email: context.email, otp: context.otp)
         if user && !user.otp_expired?
           context.message = "OTP is valid"
         else
