+++ app/services/user_authentication_service.rb
@@ -1,6 +1,6 @@
 class UserAuthenticationService
     def self.authenticate(email, password)
-      user = User.find_by(email: email)
+      user = Users::User.find_by(email: email)
       user if user&.authenticate(password)
     end
