--- test/controllers/api/v1/students_controller_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class Api::V1::StudentsControllerTest < ActionDispatch::IntegrationTest
-  # test "the truth" do
-  #   assert true
-  # end
-end
