+++ test/models/user_test.rb
@@ -1,6 +1,8 @@
 require "test_helper"
 
-class UserTest < ActiveSupport::TestCase
+class Users::UserTest < ActiveSupport::TestCase
+  fixtures :users
+
   def setup
     @user = users(:one)
   end
